# handlers/handler.go
package handlers

import (
	"context"
	"encoding/json"
	"fmt"
	"log"
	"net/http"
	"time"

	"github.com/gin-gonic/gin"
	"github.com/go-redis/redis"
	"go.mongodb.org/mongo-driver/bson"
	"go.mongodb.org/mongo-driver/bson/primitive"
	"go.mongodb.org/mongo-driver/mongo"

	models "github.com/afonsir/gin-recipes-api/models"
)

type RecipesHandler struct {
	collection  *mongo.Collection
	ctx         context.Context
	redisClient *redis.Client
}

func NewRecipesHandler(
	ctx context.Context,
	collection *mongo.Collection,
	redisClient *redis.Client,
) *RecipesHandler {
	return &RecipesHandler{
		collection:  collection,
		ctx:         ctx,
		redisClient: redisClient,
	}
}

// swagger:operation POST /recipes recipes createRecipe
//
// Create a new recipe
//
// ---
// parameters:
// - name: recipe
//   in: body
//   description: The recipe to create
//   schema:
//     type: object
//     properties:
//       name:
//         type: string
//       tags:
//         type: string
//       ingredients:
//         type: string
//       instructions:
//         type: string
//
// consumes:
// - application/json
//
// produces:
// - application/json
//
// responses:
//   '200':
//     description: Successful operation
//   '400':
//     description: Invalid input
//   '500':
//     description: Internal Server Error
func (handler *RecipesHandler) NewRecipeHandler(c *gin.Context) {
	var recipe models.Recipe

	if err := c.ShouldBindJSON(&recipe); err != nil {
		c.JSON(http.StatusBadRequest, gin.H{
			"error": err.Error(),
		})
		return
	}

	recipe.ID = primitive.NewObjectID()
	recipe.PublishedAt = time.Now()

	_, err := handler.collection.InsertOne(handler.ctx, recipe)

	if err != nil {
		fmt.Println(err)
		c.JSON(http.StatusInternalServerError, gin.H{
			"error": "Error while inserting a new recipe",
		})
		return
	}

	log.Println("Remove data from Redis")
	handler.redisClient.Del("recipes")

	c.JSON(http.StatusOK, recipe)
}

// swagger:operation GET /recipes recipes listRecipes
//
// Returns list of recipes
//
// ---
// produces:
// - application/json
//
// responses:
//   '200':
//     description: Successful operation
//   '500':
//     description: Internal Server Error
func (handler *RecipesHandler) ListRecipesHandler(c *gin.Context) {
	val, err := handler.redisClient.Get("recipes").Result()

	if err == redis.Nil {
		log.Printf("Request to MongoDB")

		cur, err := handler.collection.Find(handler.ctx, bson.M{})

		if err != nil {
			c.JSON(http.StatusInternalServerError, gin.H{
				"error": err.Error(),
			})
			return
		}

		defer cur.Close(handler.ctx)

		recipes := make([]models.Recipe, 0)

		for cur.Next(handler.ctx) {
			var recipe models.Recipe
			cur.Decode(&recipe)
			recipes = append(recipes, recipe)
		}

		data, _ := json.Marshal(recipes)
		handler.redisClient.Set("recipes", string(data), 0)

		c.JSON(http.StatusOK, recipes)

	} else if err != nil {
		c.JSON(http.StatusInternalServerError, gin.H{
			"error": err.Error(),
		})
		return

	} else {
		log.Printf("Request to Redis")

		recipes := make([]models.Recipe, 0)
		json.Unmarshal([]byte(val), &recipes)

		c.JSON(http.StatusOK, recipes)
	}
}

// swagger:operation GET /recipes/search recipes searchRecipes
//
// Search recipes by tag
//
// ---
// parameters:
// - name: tag
//   in: query
//   description: tag to filter by
//   required: false
//   type: string
//
// produces:
// - application/json
//
// responses:
//   '200':
//     description: Successful operation
//   '500':
//     description: Internal Server Error
func (handler *RecipesHandler) SearchRecipesHandler(c *gin.Context) {
	tag := c.Query("tag")
	cur, err := handler.collection.Find(handler.ctx, bson.M{"tags": tag})

	if err != nil {
		c.JSON(http.StatusInternalServerError, gin.H{
			"error": err.Error(),
		})
		return
	}

	defer cur.Close(handler.ctx)

	recipes := make([]models.Recipe, 0)

	for cur.Next(handler.ctx) {
		var recipe models.Recipe
		cur.Decode(&recipe)
		recipes = append(recipes, recipe)
	}

	c.JSON(http.StatusOK, recipes)
}

// swagger:operation PUT /recipes/{id} recipes updateRecipe
//
// Update an existing recipe
//
// ---
// parameters:
// - name: id
//   in: path
//   description: ID of the recipe
//   required: true
//   type: string
// - name: recipe
//   in: body
//   description: The recipe to update
//   schema:
//     type: object
//     properties:
//       name:
//         type: string
//       tags:
//         type: string
//       ingredients:
//         type: string
//       instructions:
//         type: string
//
// consumes:
// - application/json
//
// produces:
// - application/json
//
// responses:
//   '200':
//     description: Successful operation
//   '400':
//     description: Invalid input
//   '404':
//     description: Not found
//   '500':
//     description: Internal Server Error
func (handler *RecipesHandler) UpdateRecipesHandler(c *gin.Context) {
	id := c.Param("id")

	var recipe models.Recipe

	if err := c.ShouldBindJSON(&recipe); err != nil {
		c.JSON(http.StatusBadRequest, gin.H{
			"error": err.Error(),
		})
		return
	}

	objectId, _ := primitive.ObjectIDFromHex(id)

	result, err := handler.collection.UpdateOne(handler.ctx, bson.M{
		"_id": objectId,
	}, bson.D{primitive.E{Key: "$set", Value: bson.D{
		primitive.E{Key: "name", Value: recipe.Name},
		primitive.E{Key: "instructions", Value: recipe.Instructions},
		primitive.E{Key: "ingredients", Value: recipe.Ingredients},
		primitive.E{Key: "tags", Value: recipe.Tags},
	}}})

	if result.MatchedCount == 0 {
		c.JSON(http.StatusNotFound, gin.H{
			"error": "Recipe not found",
		})
		return
	}

	if err != nil {
		fmt.Println(err)
		c.JSON(http.StatusInternalServerError, gin.H{
			"error": err.Error(),
		})
		return
	}

	log.Println("Remove data from Redis")
	handler.redisClient.Del("recipes")

	c.JSON(http.StatusOK, gin.H{
		"message": "Recipe has been updated",
	})
}

// swagger:operation DELETE /recipes/{id} recipes deleteRecipe
//
// Delete an existing recipe
//
// ---
// parameters:
// - name: id
//   in: path
//   description: ID of the recipe
//   required: true
//   type: string
//
// produces:
// - application/json
//
// responses:
//   '200':
//     description: Successful operation
//   '404':
//     description: Not found
//   '500':
//     description: Internal Server Error
func (handler *RecipesHandler) DeleteRecipesHandler(c *gin.Context) {
	id := c.Param("id")
	objectId, _ := primitive.ObjectIDFromHex(id)

	result, err := handler.collection.DeleteOne(handler.ctx, bson.M{
		"_id": objectId,
	})

	if result.DeletedCount == 0 {
		c.JSON(http.StatusNotFound, gin.H{
			"error": "Recipe not found",
		})
		return
	}

	if err != nil {
		fmt.Println(err)
		c.JSON(http.StatusInternalServerError, gin.H{
			"error": err.Error(),
		})
		return
	}

	log.Println("Remove data from Redis")
	handler.redisClient.Del("recipes")

	c.JSON(http.StatusOK, gin.H{
		"message": "Recipe has been deleted",
	})
}